Fix #1 - First transmitted byte corrupted
When initializing the state of the transmit pin in begin(), I was
calling digitalWrite() before pinMode(). My intention was to avoid
any glitch on the pin. That works on AVR, but apparently not on
Kinetis. The pin was defaulting to LOW, which in non-inverted
voltage mode is the start bit value. That forced a framing error
on the first contiguous burst of transmitted characters.

Reversing the order to call digitalWrite after pinMode(, OUTPUT)
solves the problem.

I also noted that the Final operation was relying on the _tx_data_word
to contain a stop bit. It always does, but just to be safe I added
an explicit digitalWrite() to set the stop level between characters.