Allow control of port pins while inactive; fix flush()
Normally a software serial port releases its pins (sets them
to INPUT) when its end() method is called. Usually, if that's
an issue the programmer would just defer calling end().
With this library, though, we can only have one port active
at a time, so if there are multiple ports to serve, we need
to end() one and begin() another all the time. We need a way
to do that without letting the pins change state and cause
problems for the remote UART. We add an extension to the end()
method. You can pass SSS_RETAIN_PINS to the method to ask it
to keep control of the pins.

As a side effect, switched all digitalWrite() calls to use
digitalWriteFast() instead.

The flush() method only checked the transmit buffer count,
which goes to zero as soon as the last character is fetched
for transmission. Now it also checks the _tx_running flag,
so it won't return until after the last bit of the last
character has been completely sent through the transmit pin.

Also fixed a typo in README.md