Add two NOP tests
send_nop_with_junk() tests the UUT's ability to accept extra bytes
in the NOP command packet and not echo them back in the response.

send_nop_with_bad_crc() tests the UUT's ability to check the CRC,
but it hangs the test. Its actual purpose was to exercise the Saleae
Logic plug-in's ability to check the CRC.