Improve bad CRC sending test
Now we don't wait for a response packet (which should never come)
so we can go ahead and call this test in the normal sequence.

We add an arbitrary delay (30ms) after sending the bad CRC, just so
that the timeline (viewed on an analyzer) clearly shows the lack of
any response.